Bump cryptol submodule, adapt to GaloisInc/cryptol#1664 changes
This bumps the `cryptol` submodule to bring in the changes from
GaloisInc/cryptol#1664. This requires some minor code changes in
`cryptol-saw-core` to adapt to the `Import` data constructor now having an
`iDoc :: Maybe (Located Text)` field to represent docstrings. None of the
`Import`s that `cryptol-saw-core` generates require any docstrings, so it
suffices to instantiate them with `Nothing`.