Rework the position tracking for types.
Carry a position with each Type constructor and get rid of the extra
LType case. This is basically the same as the recent
expression/pattern changes.

However, tracking positions for types requires creating positions for
inferred types. We could just set all of these to Unknown or
PosInternal, but that isn't ultimately very helpful. The important bit
for getting useful info out is to recognize that concrete source
positions are the base case of a language object's provenance, and
therefore given a derived object it's reasonable to instead store
derivation info.

Create an additional constructor in Pos for inferred types to record
what inference was performed and the position that triggered it. Add
three cases:
   - InfFresh describes a fresh type variable created for a given
     position, such as the member type when we encounter "[]".
   - InfTerm describes a type inferred from the structure or contents
     of a term; e.g. "[]" has type List t. The position gives the
     location of the term.
   - InfContext describes a type inferred from the context or usage
     of a term; e.g. in "if e then t else f" e has type Bool.

We could add a fourth case InfUnified for types produced by
unification: it would recording the positions of both original types.
This would in some sense provide the full provenance of each inferred
type. However, for the moment at least it doesn't seem that useful; in
particular in the common case of unifying a fresh type variable with a
more specific type, recording that this happened doesn't add value. It
also gets a bit messy. So instead, add a choosePos function for
picking one of two positions to keep at unification type; it picks the
position with higher-quality information. This can be refined further
later if we so desire or find use cases.

Note that InfTerm and InfContext positions will sometimes refer to
patterns rather than terms. It does not seem worthwhile to carry that
distinction around.

In the parser, the positions of any types that appear are concrete
source positions as usual.

In the typechecker (MGU.hs), when we create a type, use one of the
forms above to describe where it came from, and use choosePos at
unification time. Also carry around unification variables as a map
from serial number to position rather than just a set of serial
numbers.

This commit is not intended to alter the functional behavior of the
typechecker.