saw-script: Fix warnings related to sbv-10.3 adding support for OpenSMT
`sbv-10.3` adds support for the OpenSMT solver
(https://verify.inf.usi.ch/opensmt). Annoyingly, it's not possible to query the
OpenSMT version number from the command line, which is the one operation that
SAW's SMT caching requires. As such, we do the bare minimum to fix warnings
related to missing pattern match cases on `OpenSMT`, but without going as far
as adding full support for OpenSMT.