Move setupArgs and friends to SAWScript.Crucible.Common
These lived in `SAWSScrupt.Crucible.LLVM.Builtins` recently, but there's not a
particularly good reason for them to live there, as they are used in both the
LLVM and JVM backends. This also paves the way for being able to use them in
the MIR backend in a subsequent commit.