CI: Upload HPC-enabled binaries to separate name
Previously, we were inadvertently uploading the HPC-enabled binaries and the
non-HPC-enabled binaries to the same name (`dist-tests-${{ matrix.os }}`). This
is very dangerous, as this risks the two artifacts clobbering each other.
Moreover, `actions/upload-artifact@v4` makes this a hard error, so we now have
no choice but to disambiguate the artifact names.