Don't fail when What4 sends us a function-based array in a result.
We can't store function-based arrays in FirstOrderValue for a number
of reasons. But we also can't do anything much useful with them, so if
we get one (unfortunately this does happen, which may or may not be
amenable to improvement in What4) treat it as an opaque blob.

Add an FOVOpaqueArray to FirstOrderValue for this. The footprint of
having this undesirable object is not large, so it's the path of least
resistance. We could in principle instead return something other than
a FirstOrderValue out of groundToFOV for this case, if we had a use
for the actual function, but since we can't do anything much with the
function because we have no idea what keys to look up, there doesn't
seem to be any point.