Document the new/upcoming issue labels. Note explicitly that label changes should include updates to the label docs.
Update the mergify config with the new issue label strings.

Drop the keep-branch-on-merge feature; have it just always remove pull
request branches at merge time. (It's not exactly difficult to put
them back if that's what you want, and it appears the one time this
feature's been used the branch hung around uselessly for years.)