Remove the short option -t for --detect-vacuity.
It turns out -t was already in use, meaning that neither -t option has
worked since it was added.

We don't think --detect-vacuity needs a short option.

Fixes the first part of #1363. (The second part is to improve the
layout of the options table so this kind of accident is harder.)