Typechecker (MGU.hs): fix unergonomic checking of function applications
When checking a function call, and it's wrong, we expect the error to
be reported as the arguments having the wrong type, not the function
being the wrong type for the arguments we provided.

Therefore, check an application by matching the argument against the
function, not the function against the arguments. In particular, we
should generate a fresh argument type and unify it with the type of
the function, _then_ unify it with the argument expression, not the
other way around.

Someone stepped on this today internally and it was confusing all
around.

Add a test case in test_type_errors.