Interpreter: rename rwTypes -> rwValueTypes, rwTypedef -> rwNamedTypes.
The former is the typing environment for values, that is, maps
variable names to their types. Calling it rwTypes is perfectly correct
but also very misleading in the sense that it's easy to think it maps
type names to their definitions. Since the key type is String there's
no obvious obstacle to this interpretation, and while I haven't found
any mistakes of this form (yet), they shouldn't be encouraged.

Meanwhile, I'm about to extend rwTypedef (which _does_ map from type
names to type information) so it doesn't only contain typedef names,
and consequently it ought to get a different name... and the natural
name is rwTypes. But that clearly won't go. So pick something else.

If anyone has better suggestions, I'm happy to hear them.

No functional change intended.