Interpreter: declare the types used by builtins, in preliminary form.
Create a table of the builtin types that aren't already special cases
in AST.hs, and initialize the named type environment with them.

(Note that some of the special cases in AST.hs probably don't need to
be special cases there and can be moved to this table, but that's for
later.)

This makes the types like LLVMType work again after the
typedef-checking changes.

Note that I haven't yet arranged to typecheck the primitives as part
of loading them. That's coming up, but needs some other fixes first.