Interpreter: remove special-case handling of syntactically top-level binds.
This corrects certain weirdnesses that were probably intended as
ease-of-use affordances for the repl but also affect scripts. In
particular:
   - monadic actions need to be in TopLevel (or in ProofScript for
     the ProofScript repl);
   - monadic actions that are polymorphic in their monad are run
     in TopLevel or ProofScript instead of being gratuitiously
     rejected;
   - the interpreter no longer wraps pieces of the bind in a
     Decl for typechecking, which is the first step on cleaning
     up the interpreter's typechecking notions (#2158).

The downside is that you can no longer directly eval non-monadic
expressions in the repl without prefixing them with "return". This is
not terrible, but we should at least consider adding repl-specific
affordances to allow that again once the code is in shape to be able
to do it reasonably.

Another oddity is that previously if you eval'd a monadic action that
was in some other monad (e.g. LLVMSetup) at the syntactic top level it
would, instead of giving a type error, just not evaluate it and then
print "<monadic>" as the result. If for some reason you really want
to do this you can also get the old behavior by prefixing the lot with
"return".

Note that all this applied to the _syntactic_ top level, and not binds
within do-blocks, whether or not they were within functions, and
regardless of whether they were in the TopLevel monad.

Include some tests and a CHANGES entry. Closes #2162.