interpreter: tidy processStmtBind more
There is no reason any more (if there ever was) to yoink the type out
of the pattern and graft a copy onto the expression as an explicit
type annotation.

Commit this separately just in case there's something subtle happening
and we end up wanting to be able to bisect to it later.