interpreter: simplify typechecker calls
The position we've been passing to checkStmt is now always just the
position _of_ the statement. The typechecker can get this on its own.

Leave a note for later that this is not really the right position to
use for the purposes the typechecker uses this position for (the monad
context the statement appears in) and that should be sorted out
better.  But it's probably not worth trying to do so until the
interfacing to the typechecker is rationalized.