tutorial: Replace SAT competition format link.
The previous link (to the 2009 format conventions for the SAT
competition) has been dead for some time. The new link takes the
reader directly to a description of the DIMACS CNF format, which is
the most relevant part of the original link for the purposes of using
SAW with external solvers.