Two tweaks to previous change.
1. Don't allow the wrong monad in a bind whose pattern is a tuple
pattern. The old saw produces a type error for that so we don't have
to allow it now.

2. For binds in the wrong monad, unify the pattern type with the full
wrong monadic type of the expression, not just the associated value
type. This assigns the variables in the pattern types in the wrong
monad, which matches both the historical behavior and also what
happens inside the interpreter during execution.

It also turns out that binds in the wrong monad with an explicit type
signature expecting the bind to work didn't typecheck in the old saw.
So we don't need to make that work as part of the workarounds either.

Add three more test cases...