typechecker: Correct the behavior of generalize.
We do need to search the value environment for mentions of unification
vars, because if we're nested inside several layers of declaration we
could be referring to unification vars coming from an outer scope, and
those shouldn't be forall-bound until we close off that scope.

We also need to search the right hand side of typedefs for the same
reason; this was previously wrong although it would be at least
moderately difficult to step on.

However, for named vars the names we should not bind are the _keys_ of
the type name environment; those are the type vars that already exist.

We don't allow random named type vars to float around any more, so
that should always exclude all named type vars we have. (I haven't
added an assertion for this, but perhaps I should have.)

Fixes #2105.