Add more typechecker tests.
This includes one test for each print, additional tests for several of
the common cases that we want to print comprehensibly, and
additionally a separate set of tests for each call to unify that is
supposed to make sure the found/expected types don't get swapped.
(Which is extremely easy to have happen, because conceptually
unification is symmetric but the found/expected reporting isn't.)

Also add some notes to the typechecker code, because it turns out some
of the typechecker messages are currently unreachable.