Rearrange the saw-script lexer.
 - Turn off the Alex wrapper. We're already doing our own version of
   half of it, and things we want to do cause type conflicts in the
   generated code we aren't using.

 - Provide our own AlexPos. AlexPos holds the current position (like
   the old lexer position type) but we don't need to carry the total
   character count in it. Stop exporting it; nothing uses it.

 - Provide our own AlexInput and alexGetByte. For the moment, panic on
   characters that don't fit in a byte; that's not quite the status
   quo but it's close enough.

 - Convert to saw-script positions (which are spans/ranges) when we
   generate tokens instead of transforming them on a second pass
   afterwards.

 - Panic on alexInputPrevChar; it is not called by the generated code.

 - Tidy various things a bit.

This commit wasn't intended to have any functional change, except
possibly for behavior on non-ASCII input that isn't currently
supported.

However, it seems that by changing the position logic I accidentally
fixed the position tracking for strings and Cryptol fragments. This
is a good thing, so run with it...

Previously the end delimiter (and for Cryptol fragments, any
whitespace before it), but not the start delimiter, was left out of
the span of the position, so e.g. {{ 3 }} would show as 4 columns wide
and the position of "abc" would end at the position of the c.

This is no longer true; now it correctly includes the whole thing.

This changes some of the reported column numbers in test_type_errors,
so update the reference outputs.