Lexer: rearrange comment handling.
Don't have the lexer match //.*; just match the // part, and extend
the same logic that handles dropping the bits between /* and */ to
also handle dropping the bits between // and EOL.

This requires materializing EOL tokens instead of treating them as
whitespace, but that costs ~nothing.

Also fix it so we generate explicit messages when there are unclosed
comments. A // comment that doesn't have an EOL before EOF generates a
warning; a /* comment without a matching */ generates an error. The
position for this error is the comment start.

(The previous behavior was that unclosed comments commented out the
EOF token and that would cause a non-specific "parse error".)

Update the lexer call sites to post the messages. There should not be
four separate lexer call sites, but that's a problem for a future
branch.

Add some tests. This includes a new test_parse_errors area akin to
the existing test_type_errors.