Set the locale explicitly to utf8 at startup, as Cryptol does.
This is not necessarily the right thing to do, but for the moment
consistency with Cryptol seems like the best approach.

I've added this in every main I can find, and in a couple cases
updated the minimum base version. (According to the docs, you need at
least base 4.5.)

This uses an internal GHC interface (apparently the only available way
to do this) but it seems like the risk of it changing gratuitously is
low, and the risk of it going away without an alternative appearing is
negligible.

The docs for the whole module suggest restricting to specific versions
of base, but that seems overkill for this particular function and
likely to cause more headaches than it would prevent.