Now move the main loop of the saw-core lexer out of the parser monad.
Put it instead in the Either monad (with () on the unused error side)
so as to not have to change all the do-notation.

(These changes are small, but this code is delicate, not well tested,
and confusingly written in continuation-passing style, so I want it to
all be readily bisectable later.)

No functional change intended.