saw-core lexer: make comment state machine not hang on unclosed comments.
Add some tests.
Fixes #1225.

Note that the error messages are complete rubbish so far (they seem to
arise from default Show instances), but that's a problem for later.

We also could have done this earlier but now the change is obviously
correct.

(except that it prints the end-of-file position with the error message;
we would like to print where the open comment is, but that requires
rearranging the state machine, and that's coming next)