saw-core lexer: fix the interaction of block and line comments.
As noted in #1225, it should not be possible for a line-comment to
comment out the end of a block comment.

Instead of matching --.*, materialize tokens for -- and \n and run
both those and the {- -} tokens through a single state machine to drop
comments.

(This is essentially the same treatment as given the saw-script lexer
a few commits back.)

Update the tests: the test1225 test for this specific point now works
properly. Also the more complex state machine lets us track the start
location of the current block comment so we can print that when it's
unclosed, as mentioned would be desirable in the last commit. This
updates the corelexer002 test output.