saw-core parser: simplify the lexer call
The normal way to do this is %lexer { lexer >>= }, not %lexer { lexer }.
This allows "lexer" to have a reasonable signature.

(Although it still lives in the parser monad, so it's still not just a
direct call into the lexer itself.)

Hat tip to Eric Mertens who pointed this out.

No functional change intended.