Track lifecycle (deprecated/experimental) settings in the interpreter.
Previously, when you did enable_experimental or enable_deprecated, it
would go look at the big table of builtins, find the ones so tagged,
and stuff those into the current environment on the fly.

Now, we load them all at startup and inspect the state when they're
referenced, and enable_* just changes what's visible.

This allows various better behavior like not suggesting
enable_experimental for every unbound variable error. It also will
allow marking things as warn-deprecated as well as hide-deprecated.
(We've wanted that for a while now.)

Not all such improvements are in this changeset just yet though. The
warn-deprecated changes are coming in the next commit.

In addition to adding the lifecycle setting to each symbol in the
value environment, this commit also folds the previously separate
value-value (variable name -> value) and value-type (variable name ->
type) environments together, and adjusts the names of these fields of
TopLevelRW to be less confusing. Having one environment for
value-level variables instead of two makes it impossible to add
variables with values but no types, or variables with types but no
values, both of which make no sense and shouldn't be allowed. It was
possible in the past for these environments to get out of sync and
lead to strangeness. On the code path that allowed that, insert a
bottom type instead. This will get cleaned up further shortly.

This commit also includes improvements to :search so that experimental
and deprecated results are handled nicely.