rust-tutorial: Subtract 1 from all headings after the first.
This temporarily makes this Markdown violate the "thou shalt only have
one <h1>" rule, however: Logically, we want this part of the SAW
documentation to be split into chapters, similar to the LLVM/Java
tutorial.
Before, this part of the documentation rendered to PDF (via Sphinx) as
one giant chapter.