tutorial: Remove PDF (+ generation) and split up tutorial.md.
Not in that order :)

In an effort to start standardizing the Markdown (e.g. removing
multiple top-level headers), this commit primarily restructures the
LLVM/Java tutorial into multiple source files.

This additionally makes individual parts of the tutorial easier to
maintain.
A similar change will be implemented for the manual and Rust tutorial
for the same reason.

This restructuring naturally broke the original PDF generation;
fortunately, the eventual introduction of Sphinx will give us both
HTML and PDF generation for free, and this generation strategy is no
longer necessary to maintain.

Note that this commit does _not_ perform any changes to the content of
the tutorial or indeed any other structural property of the Markdown.
Additional changes (e.g. standardizing the Markdown syntax used) will
be the subject of future commits.