Rename internal -> development.
These documents may not be intended for most users, but they ought to
be available to read in a nicer format than Markdown files in the
repository.
Since we control which documents are rendered to PDF (i.e. when we
`make latexpdf`) explicitly, these developer documents will only be
rendered as part of the HTML (that will eventually be our GitHub Pages
site for saw-script).

This change reflects that these are documents related to the
development of SAW (rather than generically 'internal').
Additional restructuring will follow in additional commits.