releasing: Fix link to `CHANGES.md`.
And by 'fix', I mean: Add a document containing a verbatim rendering
of the 'real' `CHANGES.md` at the SAWScript repository root, and link
to that.

An alternative that was considered that would have resulted in the
'real' `CHANGES.md` being rendered directly was the `{include}`
directive.
Not only is it the case that "the include directive represents a
potential security hole"[^1], Sphinx will attempt (and fail) to
resolve the links within, which will fail fantastically since most of
them are outside of the doc/ root.

[^1]:
https://docutils.sourceforge.io/docs/ref/rst/directives.html#include