Update CONTRIBUTING to reflect the current state of affairs.
Add doc/developer/developer.md as a general spill area for things we
might put in CONTRIBUTING but that a new developer doesn't actually
need to see right away.

Also move the issue label documentation to its own file as it turns
out it's rather large (about a quarter of the total line count in
these files...)

This adds a bunch of stuff from our processes and practices
discussions last summer, but not really everything yet; there'll want
to be another go, and possibly another reorganization first. We
should see how we feel about the way this is laid out in another
couple months.

Some specific bits:
   - update the repository layout summary -- it was quite out of date;
   - mention a bunch more tools you want if you're developing;
   - mention what4-solvers;
   - mention that on MacOS you need a real JDK and not the placeholder
     the system ships with;
   - link to a bunch of the open issues related to these topics.