Add indices for all SAW documentation.
In Sphinx, document hierarchy is determined by the toctree directive.
That is the most important function of these new files -- defining the
structure of each SAW 'book', and giving an overall structure to the
documentation as a whole.

In particular, these indices result in the following document
tree:

SAW Documentation/
├─ LLVM/Java Verification with SAW/
│  ├─ Introduction
│  ├─ ...
├─ Rust Verification with SAW/
├─ SAW User Manual/
├─ SAW Developer Documentation/
│  ├─ SAWCore External Format/
│  ├─ Limitations
│  ├─ Releasing

Additional 'books' can be added to SAW Documentation; each will have
its own toctree(s) defining its internal structure in an index.md file
(see doc/index.md, the "root" of the eventual Sphinx project, to
better understand how this might be done).