saw-lexer: Add a simple pygments lexer for SAWScript.
This is (mostly) based on the SAWScript Alex lexer; because pygments
is quite a bit more simplistic, we have to be more careful when
defining its states.

I did not test this beyond the context of inline SAWScript in the
documentation, and don't really plan on testing further at this time.
As long as we have a passably-functional way to lex SAWScript for
documentation, we should be fine.

Ideally, this code wouldn't live here - it would be added to pygments.
That said, the accompanying pyproject.toml means that, when installed,
saw-lexer acts as a seamless extension of pygments; we won't have to
do anything special when running Sphinx to tell it about this lexer.
Additionally, if/when the lexer is added to the official pygments
release, the changes required to the documentation build will be
extremely small.

Note in particular that the matching for builtins is lackluster, in
that it is both liberal (based on basic prefix-matching) and
incomplete (I am certain I missed things).
I don't think it is anywhere close to our top priority, so for now I'm
happy with what I saw in generated (HTML/PDF) documentation.