Add a Script that creates a Python virtual environment for SAW docs.
Most documentation authors need not be concerned with the minutiae of
the Python setup; while it is simple/typical of other Python projects,
it's annoying to have to think about if your job is to be writing or
maintaining documentation.

This commit adds a script to help.

First and foremost, if python3 is not available, it bails right away.

Then, we handle an existing .venv: If it's ended up in some kind of
botched state without an activation script, we run python3 -m venv
.venv to try to fix it.
Note that this is not robust to an ill-formed but existing
.venv/bin/activate; that really didn't seem worth the effort here.

We then activate this mystery environment, and check whether or not it
has the proper dependencies installed (as given by requirements.txt,
see 71a2d4a71 for more detail).
If any are "not installed" (according to pip freeze output), we simply
install from requirements.txt to resolve the issue.

Finally, if no .venv was found to begin with: We combine the essence
of these cases to create the environment.

There is probably more error-handling than necessary in this script,
but it's nice to have something that handles most of the likely
scenarios one may find themself in.
Additionally, with the exception of the case described above (where
.venv/bin/activate does exist, but isn't the right thing), this script
has the property that, after being run, a ready-to-go Python virtual
environment is available at the .venv directory, i.e., sourcing
.venv/bin/activate will get you into a state ready to build SAW
documentation (or hack on the documentation build process itself).