Commit PDF renderings of the tutorials and manual.
At long last, we've run the build!

Just kidding, it has been run many times to get us to this point :)
Now is just the right time in the 'story' to commit these.

@sauclovian-g suggested a CI check that compares PDF renders built in
CI to the checked-out renders.
Before that, I would advocate to remove the PDFs from the repository
entirely, and provide them instead as 'releases' attached to commits
to the default branch.

That said, there are problems with trying to compare PDFs anyway,
especially these TeX-generated PDFs that aren't the most portable
things to begin with.
Sphinx _does_ attempt to avoid re-building PDFs when working locally,
but it seems to do so based on the state of the _build directory (a
simple test where one of the PDFs is removed from that location and
re-built confirms this).

In any case, to continue providing the PDFs as before, this commit
adds the result of running ./make_docs.sh latexpdf on my machine.