typechecker: don't drop updated patterns
Typechecking updates some the AST elements with additional
information, including Pattern, and some (but not all) of the code
paths were discarding the updated Pattern. It's not clear if this was
actually causing any real problems, but it's wrong and it seemed like
a good idea to fix it before tinkering with patterns further.

(which then didn't happen because the original approach to the
upcoming stuff didn't work out, but anyway...)