rust-verification-with-saw: Replace missing '`' in Cryptol type var.
Because it created an ill-formed inline code block, I removed a backtick from a
Cryptol type variable in an earlier commit.
This was a hasty decision, and incorrect!

I also learned from this experience that it's possible to use double backticks
in Markdown when your inline code itself contains backticks.
Neat, and fits nicely given that adding a third naturally creates a code block
'fence' that renders as a block element.