saw-lexer: Remove special 'Name.Builtin' handling.
First, my comment was inaccurate -- there is a big table of builtins in the
interpreter (thanks @sauclovian-g for pointing this out).
As he rightfully noted, one day we will like to have these / their docstrings
generate an appendix for the user manual.

Second, other than looking a bit fancier in the render, there's not too much
reason to give special treatment to the builtin functions and how they render;
so, since we already plan to generate documentation of the builtins eventually
anyway, we can let them be lexed as regular identifiers (and in the future add
references to the appendix).