Remove CHANGES mirror from developer docs and restore link text.
We're going to hide developer docs from Sphinx anyway, but this essentially puts
things back to how they were before this PR (with the addition of an extra '..'
to get to the right place.

Essentially, this change reflects our desire for doc/developer/releasing.md to
be readable on GH, and properly link to the change log at the repository root.