Hide SAW developer documentation.
As originally planned, this commit hides the development/ directory from the
Sphinx build, and removes all links to it from the document tree.

See https://github.com/GaloisInc/saw-script/pull/2202#discussion_r1933024199 for
relevant conversation on the matter of publishing developer documentation.
TL;DR: For now, there is nothing we want to publish (extcore is not something
intended for broader use than loading/storing SAW Terms, the release process is
only relevant to developers, and the limitations document is old / needs review
before being added to user-facing documentation, at which time it would likely
be best added to the manual instead of as a separate resource).