Makefile: Re-implement the behaviors of the old make_docs.sh.
Thank you to @sauclovian-g for suggesting this rearrangement - not only is it
"more correct", it resulted in smaller, more reusable scripts and a more
intuitive build process.
That's great!

First, we explicitly list SUPPORTED_DOCS (that is, the targets Sphinx's old
Makefile understood that we wish to continue supporting).
This allows us to (1) always show up-to-date help text documenting what someone
can run to build the docs, (2) easily add additional format support, and (3)
replace the old catch-all target with a more helpful suggestion given these
changes.

We replace Sphinx's help with a series of messages listing available targets,
but still leave Sphinx's original help available (this is mostly for developers
looking to add support for additional Sphinx targets) via target sphinx-help.

Targets setup-env and package-code are methods of executing the scripts in
scripts/.
These are also the dependencies of the document-rendering targets, and anything
related to running Sphinx.

mostlyclean and clean respectively clean up the documentation build (including
package code examples) and all of that plus the Python environment.

install-pdf copies rendered PDFs from the Sphinx output directory to doc/pdfs,
building them if necessary.