Generate PDFs in a (hopefully-reproducible) way.
To eventually validate that the 'correct' PDFs have been checked in, this commit
adds a mechanism to control the reproduction of SAW documentation PDFs: The
SOURCE_DATE_EPOCH environment variable (see
https://reproducible-builds.org/docs/source-date-epoch/).

We set this in a .env file, and use if in the Makefile on documentation builds
so the last edit time is consistent.

For now, this value is set to (GMT) Tuesday, April 1, 2025 12:00:00 AM, but for
releases should be set to an epoch timestamp corresponding to the date of
release (for example).

Release instructions for the repository have been updated to reflect the need to
update this variable at release time (and correspondingly, the documentation
PDFs themselves).