sphinx-download-dir: Sphinx role to download directories.
This Python package implements a new Sphinx role, download-dir, that can be used
to indicate that a referenced directory should be archived and made
downloadable.

For example: If we write {download-dir}`/my/cool/dir`, the role attempts to
create `$SOURCEDIR/my/cool/dir.tar.gz` from the contents of
`$SOURCEDIR/my/cool/dir`, where $SOURCEDIR is the Sphinx project root (for us,
`saw-script/doc/`).

In short, this means that you can only create downloads out of directories
somewhere under in `doc/`, including `doc/` itself.

As is Sphinx's convention, then, paths starting with `/` indicate that the
path is rooted at $SOURCEDIR, and relative paths are relative to the document
where the role was used.

The role also supports
{download-dir}`links with explicit titles <thing/to/download>`; when no title is
provided, the link will be the name of the archive (i.e. the stem of the target
directory + ".tar.gz").

This is an improvement over our current solution, which is inflexible (requiring
that only specific directories be used as download targets) and doesn't fully
utilize Sphinx's/docutils's extensible role framework (instead using scripts).