Remove scripts/package_code.sh (and Makefile uses of the same).
Given that we now use sphinx-download-dir (which handles the archiving of code
samples during the Sphinx build), we have no need for the packaging script /
make target.

Furthermore, we make the mostlyclean target interactive in its deletion of
tar.gz archives, to prevent unintentionally deleting a checked-in archive that
is part of a sample directory (the Rust tutorial is a case like this).

This is mostly necessary since we no longer force the sample directory / archive
to have a specific name, and simply attempt to clean up _all_ tar.gz archives in
and under `doc/`.